Add GMTL_WARN_UNUSED_RESULT, use on changed methods.
Some methods used to modify their first argument and didn't require
that the return value be used, but changes to the math library
to improve generated code have made that no longer be the case
and the return value must be used.

The support for this in Visual Studio requires that someone be
using the static analysis tools, but clang handles this fine
out of the box.